% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_test.R
\name{bf_test}
\alias{bf_test}
\title{Perform the Brown-Forsythe Test for Equality of Variance}
\usage{
bf_test(model, num_groups = 2, resid_type = "jackknife", plot_graph = TRUE)
}
\arguments{
\item{model}{A model of type lm or glm.}

\item{num_groups}{The number of groups to be used. The default is 2.}

\item{resid_type}{The type of residuals. The default is "jackknife", but
supports "raw", "standard", and "pearson".}

\item{plot_graph}{Whether to return a plot or not. The default is TRUE.}
}
\value{
This function returns an ANOVA table and, if requested a plot of the
residuals
}
\description{
When given a model, this will break up into groups of equal size with a
default of 2 perform a Brown Forsythe test. By default, the test will be
performed with the jackknife residuals.
}
\examples{
mod <- lm(mpg ~ disp, data = mtcars)
bf_test(mod)

}
