% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births}
\alias{births}
\title{births Data Set}
\format{
A tibble with 2 variables:
\describe{
\item{prenatal}{A factor indicating the prenatal care the mother received:
Adequate, Inadequate, or Intensive.}
\item{type}{A factor indicating the classification of the birth: "Preterm
(induced or cesarean)", "Preterm (without procedures)", and
"Term or post-term"}
}

A tibble with 4 variables:
\describe{
\item{sex}{A factor indicating the sex of the person: Female or Male.}
\item{actual}{The person's actual weight.}
\item{ideal}{The person's ideal weight.}
\item{diff}{The difference between the person's actual weight and their
ideal weight. Negative values indicate that the person weighs less than
what they consider ideal.}
}
}
\usage{
births

births
}
\description{
Data from 1995-1997 for a study hat examined pregnancies that resulted in the
birth of twins. Births were classified as preterm with intervention (induced
labor or cesarean), preterm without procedures, or term/post-term.
Researchers also classified the pregnancies by the level of prenatal medical
care the mother received (inadequate, adequate, or intensive). The data set
consists of 278 cases (rows) with two columns indicating the level of
prenatal care and type of birth for each set of twins.

This data set contains weight information on 182 people (119 females and 63
males). Actual weights, ideal weights, and the difference between them are
recorded.
}
\keyword{datasets}
