% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_capture.R
\name{ci_capture}
\alias{ci_capture}
\title{Computes many confidence intervals and returns how many capture the true mean}
\usage{
ci_capture(n, mu, sigma, conf_level = 95, ci_type, n_ints, plot_graph = TRUE)
}
\arguments{
\item{n}{Sample size of each sample.}

\item{mu}{The true population mean.}

\item{conf_level}{The confidence level. By default, this it 95 for 95\%
intervals.}

\item{ci_type}{The type of confidence interval to create. "z" for a
z-interval and "t" for a t-interval.}

\item{n_ints}{The number of intervals to create.}

\item{plot_graph}{Whether to return a plot or not. The default is TRUE.}
}
\value{
This function returns a graph of all the intervals, if requested, and
the number of intervals that contained the true mean.
}
\description{
This function reads in a given sample size, mean mu, standard deviation
sigma, confidence coefficient, confidence interval type ("z" or "t"),
and the number of simulated samples desired, and returns a count of how
many of the corresponding confidence intervals captured the true mean mu.
}
\examples{
ci_capture(10, 10, 5, 95, "t", 500)

}
