% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_summary.R
\name{multinom_summary}
\alias{multinom_summary}
\title{Outputs estimates, standard errors, test statistic, and p-values for a
multinomial logistic regression object from the nnet library.}
\usage{
multinom_summary(model, which = NULL)
}
\arguments{
\item{model}{An object of type "multinom" "nnet"}
}
\value{
This function returns a list of tibbles with the coefficient
estimates, the standard errors, the test statistics, and the p-values for
the models comparing each category to the baseline. The AIC and deviance are
also given as the last elements in the list.
}
\description{
When given an nnet multinom object, this function will compute the test
statistic and p-value for each term in the models and return a tibble with
the coefficients, standard errors, test statistic, and p-values in a list
with one list element per model.
}
\examples{
mod <- mtcars |> 
  mutate(carb = factor(carb)) |>
  multinom(carb ~ mpg + disp + hp, data = _)
multinom_summary(mod)

}
