% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_test.R
\name{norm_test}
\alias{norm_test}
\title{Performs either the Shapiro-Francia or the Shapiro-Wilk normality test}
\usage{
norm_test(model, resid_type = "raw", test = "default", plot_graph = TRUE)
}
\arguments{
\item{model}{A model of type lm or glm.}

\item{resid_type}{The type of residuals. The default is "raw", but
supports "jackknife", and "standard" as well.}

\item{test}{The type of test that should be performed. Options are "default",
"sf", and "sw". "sf" performs the Shapiro-Francia test, "sw" performs the
Shapiro-Wilk, and "defualt" performs the Shapiro-Francia is n > 30 or
performs the Shapiro-Wilk if  n <= 30.}

\item{plot_graph}{Whether to return a plot or not. The default is TRUE.}
}
\value{
This function returns a test statistic, and p-value and, if requested
a plot of the residuals
}
\description{
When given a model, this will perfrom either the Shapiro-Francia or the
Shapiro-Wilk normality test depending user request or based on sample size.
Both of these tests have a null hypothesis that the residuals are normally
distributed and an alternative that the residuals are not normally
distributed.
}
\examples{
mod <- lm(mpg ~ disp, data = mtcars)
bf_test(mod)

}
