% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_levene.R
\name{test_levene}
\alias{test_levene}
\title{Performs Levene's test many times to investigate standard deviations needed
to reject that the true SDs are equal}
\usage{
test_levene(n, num_sims)
}
\arguments{
\item{n}{Sample size of each group.}

\item{num_sims}{The number of Levene's tests that should be performed.}
}
\value{
This function returns a list that contain the standard deviations for
each group when we reject the null hypothesis in Levene's test, the p-values
for each time we rejected the null hypothesis, the number of samples, the
ratio of largest SD to smallest SD for each time we rejected the null, and
the number of times we rejected the null.
}
\description{
This function performs Leven's test for the equality of variance in the ANOVA
setting many times for three randomly generated normal samples when given the
sample size and records how often the null is rejected, what the standard
deviations for each group is when the null isrejected, the p-values when the
null was rejected, and the ratio of largest SD to smallest SD for each time
the null was rejected.
}
\examples{
test_levene(10, 1000)

}
